refactor: move auth to service
