+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:aimimi/models/user.dart';
-import 'package:aimimi/providers/auth.dart';
+import 'package:aimimi/services/auth_service.dart';
 import 'package:aimimi/services/goal_service.dart';
 import 'package:aimimi/constants/styles.dart';
 import 'package:aimimi/views/login_view.dart';
