+++ lib/views/login_view.dart
@@ -1,4 +1,4 @@
-import 'package:aimimi/providers/auth.dart';
+import 'package:aimimi/services/auth_service.dart';
 import 'package:aimimi/constants/styles.dart';
 import 'package:aimimi/views/signup_view.dart';
 import 'package:aimimi/widgets/background_painter.dart';
