+++ lib/views/main_view.dart
@@ -1,7 +1,7 @@
 import 'package:aimimi/constants/styles.dart';
 import 'package:aimimi/models/goal.dart';
 import 'package:aimimi/models/user.dart';
-import 'package:aimimi/providers/auth.dart';
+import 'package:aimimi/services/auth_service.dart';
 import 'package:aimimi/services/goal_service.dart';
 import 'package:aimimi/views/shares_view.dart';
 import 'package:aimimi/views/activity_view.dart';
