+++ lib/views/profile_view.dart
@@ -1,4 +1,4 @@
-import 'package:aimimi/providers/auth.dart';
+import 'package:aimimi/services/auth_service.dart';
 import 'package:aimimi/constants/styles.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
