+++ lib/views/signup_view.dart
@@ -1,5 +1,5 @@
 import 'package:aimimi/constants/styles.dart';
-import 'package:aimimi/providers/auth.dart';
+import 'package:aimimi/services/auth_service.dart';
 import 'package:aimimi/views/login_view.dart';
 import 'package:aimimi/widgets/background_painter.dart';
 import 'package:flutter/material.dart';
